RUM-593: Fix unit tests

+++ features/dd-sdk-android-rum/src/test/kotlin/com/datadog/android/telemetry/assertj/TelemetryConfigurationEventAssert.kt
@@ -290,7 +290,7 @@ internal class TelemetryConfigurationEventAssert(actual: TelemetryConfigurationE
         return this
     }
 
-    fun hasOkhttpInterceptorSamplingRate(expected: Float?): TelemetryConfigurationEventAssert {
+    fun hasOkhttpInterceptorSamplingRate(expected: Long?): TelemetryConfigurationEventAssert {
         assertThat(actual.telemetry.configuration.traceSampleRate)
             .overridingErrorMessage(
                 "Expected event data to have telemetry.configuration.traceSampleRate" +
